Add mirror_checker for SMAPIv3 migrate
There is a mirror_checker/tapdisk_watchdog for SMAPIv1 that periodically
checks the status of the mirror and sends an update if it detects a
failure.

Implement something similar for SMAPIv3 mirror, although this
check happens for a shorter period of time compared to the SMAPIv1
tapdisk_watchdog because the `Data.stat` call will stop working once the
VM is paused, and currently we have no easy way to terminate this mirror
checker just before the VM is paused (in xenopsd). So only do this check
whilst the mirror syncing is in progress, i.e. when we are copying over
the existing disk content.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>