Cancel the watchdog before deactivating SMAPIv1 VDI
Previously the tapdisk watchdog in SMAPIv1 mirroring was cancelled in
the `post_deactivate_hook`, but at that point the VDI has already been
deactivated, and hence the mirror would have been terminated.
Additionally, the last time the stats is retrieved is in
`pre_deactivate_hook`, so do this cancelling after the last stats
retrival.

Note that SMAPIv3 mirror does not have a watchdog due to the limitations
of the mirror job auto cancel after guest pause, so instead the mirror
checking is only done whilst the mirror syncing (i.e. copying existing
disk content) is in progress.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>