Add post_deactivate_hook to storage_mux for SMAPIv3
This is a bit of a layering violation as storage_mux should not care
about the version of SMAPI the SR is, nor should it be responsible for
calling hook functions. But as there is no way for xapi-storage-script
to invoke code in xapi (which would also be a layering violation if it
was possible), and smapiv1_wrapper has special state tracking logic for
determining whether the hook should be called. Leave the hook here for
now.

Note the pre_deactivate_hook is not called as currently that remains a
noop for SMAPIv3. And as we do not support VM shutdown during outbound
SXM for SMAPIv3 anyway, leave a hack in the storage_mux for now until we
have a plan on how to support that.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>