[prepare]: call Share.merge around ensure_utf8_xml calls
This needs to be called around the same places that `ensure_utf8_xml` and `UTF8_XML.is_valid` gets called:
those are the places where new strings can enter the system.

Now that calling Share.merge is thread-safe also call it when writing new fields,
not just when creating the row.
For now the DB lock is still held when this function is called, but future refactoring might move code around.

For maps and sets we only call this when new keys get added (except in create_row, where we need to process all elements)

Signed-off-by: Edwin Török <edwin.torok@cloud.com>