CA-408492: Add back receive_start2 for backwards compat
This function was introduced during the development of inbound SXM for
SMAPIv3, but was used in SMAPIv1 migrations as well as SMAPIv1 ->
SMAPIv3 migration, so add this back to SMAPIv2 for backwards
compatibility.

Migrations from xapi-(25.3.0 -- 25.17.0) would otherwise be broken if
they were migrating to the latest version of xapi. Yangtze, however, is
not affected.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>