CP-53642: change default NUMA placement policy to best-effort
We've seen that using the policy can be up to 10% faster than using any is some
workflows, while not observing workflows that were negatively affected. The
policy per VM can always be change if need be.

Note that currently sometime the best-effort falls back to the same behaviour,
especially when restarting on starting more than one VM at a time. This needs
xen patches to be fixed:
https://lore.kernel.org/xen-devel/20250314172502.53498-1-alejandro.vallejo@cloud.com/T/#ma1246e352ea3cce71c7ddc26d1329a368548b3b2

Now the deprecated numa-placement configuration option for xenopsd does nothing.
It was exclusively used to enable Best_effort, since now it's the default,
there's no point in setting the option. It's value depends on whether the
default option is best_effort or not, as per the spec.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>