CP-53642: change default NUMA placement policy to best-effort
We've seen that using the policy can be up to 10% faster than using any is some
workflows, while not observing workflows that were negatively affected. The
policy per VM can always be change if need be.

Note that currently sometime the best-effort falls back to the same behaviour,
especially when restarting on starting more than one VM at a time. This needs
xen patches to be fixed:
https://lore.kernel.org/xen-devel/20250314172502.53498-1-alejandro.vallejo@cloud.com/T/#ma1246e352ea3cce71c7ddc26d1329a368548b3b2

Also fix the legacy numa-placement configuration option for xenopsd. It was
always deciding the setting, even when not used, not it only takes effect when
it's present, otherwise it leaves the default option untouched.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>