CP-53719: Update monitor and collecting list for host network devices
With interface-rename functionality, the host network devices are
ordered and renamed like "eth<N>". So they can be filtered by
hard-coded "eth" and added into the networkd monitor list and rrdd
collecting list.

While the interface-rename is being replaced by the ordering function
in networkd. Unlike the interface-rename, the networkd will not rename
the host network devices anymore. Instead, the naming of them are left
to the built-in systemd and Kernel.

In both cases, the networkd's Interface.get_interface_positions can
return the managed (ordered) host network devices: when the
interface-rename functionality is still working, it returns the list of
"eth<N>"; when networkd takes the place to perform the ordering, it
returns the names generated by systemd and Kernel.

Signed-off-by: Ming Lu <ming.lu@cloud.com>