SXM: Keep previous http handler for back-compat
In commit 2eff6ab722f631e4345a0cb14e3856094dbceeb6,
the http handler was renamed to add an "import" in the url, but we need
to keep the previous one for backwards compatability. This is so that
previous versions of sparse_dd in XS 8 can migrate to the latest one.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>