CA-409431: Use an Observer forwarder for xapi-storage-script
Currently, xapi-storage-script uses the presence/absence of a smapi
observer config file to determine whether it should create traces. This
only happens on startup which means smapiv3 traces will often not be
created when they should be.

This commit updates the Smapi Observer forwarder to use an RPC client to
send messages to xapi-storage-script, updating it on any relevant
changes to the Observer.

Signed-off-by: Steven Woods <steven.woods@cloud.com>