xenopsd: expose a best-effort mode that set the hard affinity mask (CP-54234)
This allows xapi to "hard-pin" instead of "soft-pin". This is useful to
restrict the scheduler to a single numa node, instead of letting it to move the
vpus across nodes. This must be used with care because the effect of
overprovisioning CPUs in this mode is unknown and will probalby have undesired
effects.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>