CA-409949 CA-408048 XSI-1912 remove unabailable SM plugin by ref
When we have multiple SM plugins in XAPI for the same type (which
happens only because of past problems) and want to remove the obsolete
one, do this iby  reference. The code so far was assuming only one per
type and looked up the reference by name which was not unique and hence
could end up removing the wrong SM entry.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>