XSI-1918: Host can not join pool after enable external auth
commit 6ac37d770 add localhost name to /etc/hosts
There problem is, instead of appending FQDN to the end of the line,
the entry should in following format:
IP FQDN [alias1] [alias2]

Besides, another non-regression is found.
The host need to change its static IP and DNS to join AD,
However, then the Host record is not send to DNS, cause the pool
member can not resovle each other.

To resovle the issue, a task is scheduled to sync the hostname
to DNS.

Signed-off-by: Lin Liu <Lin.Liu01@cloud.com>