xapi_vif: Guarantee the device parameter is an unsigned decimal integer
This has been always true as xapi will call valid_device on VIF creation to
make sure device is an integer, but the datamodel type of 'device' is string,
without any such guarantees.

Specify the guarantee in the documentation and make the check stricter
(int_of_string will accept "0x9fe" as an integer, for example), making sure
that the device is specifically a decimal unsigned integer. allowed_VIF_devices
has already enforced the unsigned decimal integer limitation on compliant clients.

This could be helpful in ensuring that the clients will always be right in
sorting devices as numbers, not as strings (so that "2" follows "1" instead of
the string order of "1"->"10"->"11", etc.).

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>