xapi-cli-server: Add some of the missing parameters to cli_frontend
There were several optional boolean parameters that were checked and used in
cli_operations but were not included in cli_frontend (therefore would not be
shown in suggestions or 'xe help command').

Add these to cli_frontend. is_unique is a ... unique case because it does not
follow the style of the CLI parameters (which use dashes, not underscores, to
separate words), so add 'is-unique' to cli_frontend but handle both in
cli_operations.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>