Rename RRD metric "runnable" to "runnable_any"
The renaming is done to resolve one problem with rrd2csv: rrd2csv
selected both RRD1("runnable") and RRD2("runnable_vcpus") when the RRD1
name is used:

> rrd2csv AVERAGE:vm:<vm-uuid>:runnable
> timestamp, AVERAGE:vm:<vm-uuid>:runnable, AVERAGE:vm:<vm-uuid>:runnable_vcpus

This is because "runnable" is a prefix of "runnable_vcpus".

After the renaming, with rrd2csv:
* can select only RRD1 if we use:
  rrd2csv AVERAGE:vm:<vm-uuid>:runnable_any
* can select only RRD2 if we use:
  rrd2csv AVERAGE:vm:<vm-uuid>:runnable_vcpus
* can select both RRD1 and RRD2 if we use:
  rrd2csv AVERAGE:vm:<vm-uuid>:runnable

And the renaming also makes it clearer as the 'runnable' metric is % of
time that at least one vCPU of the domain is in the runnable state.

Signed-off-by: Gang Ji <gang.ji@cloud.com>