CA-412636: hostname changed to localhost with static IP and reboot
According to
https://www.freedesktop.org/software/systemd/man/latest/hostname.html
Systemd set hostname with following sequence
- kernel parameter, systemd.hostname
- static hostname in /etc/hostname
- transient hostname like DHCP
- localhost at systemd compile time

Once the host is configured with static IP and reboot, it would
just lost its hostname as no DHCP or static IP available.

However, the hostname is critical to AD function as it construct the
machine account.

The hostname should be persisted as static name during joining AD,
this is also what PBIS does.

Note: the static hostname is not cleaned during domain leave.
This is by intention to avoid losing hostname after reboot with static IP

cba2f1d5e tried to resovle the issue and update /etc/resolv.conf
However, /etc/resolv.conf does not help and conflict with
xcp-networkd, as networkd override the configure every 5 minutes
Here we just revert the resolv.conf update. Other parts of that commit
can still benifit as it push the hostname to DNS

Signed-off-by: Lin Liu <Lin.Liu01@cloud.com>