xenopsd-xc: do not try keep track of free memory when planning NUMA nodes (CA-411684)
Free memory is now properly accounted for because the memory pages are claimed
within the NUMA mutex, so there's no need to have double tracking.

On top of that, this code never increased the free memory, which means that it
always reached a point where it was impossible to allocate a domain into a
single numa node.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>