CA-408552: 1/3 Improve bootstrom performance by save db ops
events_from_xenopsd thread is critical as it sync up VM status
in case of bootstorm, this thread is flood as lots of events comes
from xenopsd waiting for process.

During processing of the events, VM/VDI/VBD update_allowed_operations
will be called to refresh the allowed operations.

However, for each ops (start/suspend,etc) for the same object(VM),
the object info is always the same no matter what the ops is.
Thus, it is not necessary to query the object information over and
over again.

Disclosure is used to resovle the issue. Query once and the disclosure
will just remember the query result.

The performance test for starting 500 VM on 4 hosts improve around 10%
performance for both XS8 and XS9

This commit just introduce disclosure and caller call the disclosure
instead of the original function

Signed-off-by: Lin Liu <Lin.Liu01@cloud.com>