CP-308450 Use the position in new network name_label
Do not use the device name in the new created network name_label
if the device has a position. Given the fact that device names
may be different on different hosts of the pool, even they are of
the same position. So it is more proper to just use the position
in the network name_label.
Also this will be good for interface name change scenario as this
change will not reflect in this default network name_label.

Signed-off-by: Changlei Li <changlei.li@cloud.com>