xenopsd: set xen-platform-pci-bar-uc key in xenstore
This patch add a new parameter named 'xen-platform-pci-bar-uc' who has a
default value of 'true' to keep the default behavior of hvmloader.
Putting 'false' to this parameter will tell xenopsd to add a xenstore
key of '0' in:
'/local/domain/<domid>/hvmloader/pci/xen-platform-pci-bar-uc'.
Only this key set to 0 will change the behavior of hvmloader.

This changeset is link to this xen commit:
https://xenbits.xen.org/gitweb/?p=xen.git;a=commit;h=22650d6054625be10172fe0c78b9cadd1a39bd63

Signed-off-by: Anthoine Bourgeois <anthoine.bourgeois@vates.tech>