rrdp-squeezed: Don't collect metrics from a domain with missing counters
Previously if only some of the values was missing, the arithmetic operations
where done and they were taken into account into the host total.

When a domain's values are missing, they are not taken into the sum.

Also prepare the code to add per-VM metrics by using shared infrastructure and
using a record.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>