rrdp-squeezed: generate per-vm memory target datasources
The daemon was already watching the vm memory targets, so it makes sense to
make it generate the datasources directly as well.

This uses the shared xenctrl infrastructure that's needed to fetch the domains'
uuids.

This means that xcp-rrdd does not need to collect this anymore, and xenopsd
does not need to send it to xcp-rrdd.

The function in the idl is safe to delete because the only user was xenopsd,
and it was well-protected against errors. This means that if an old version of
xenopsd tries to call the function while xcp-rrdd has already been updated, it
won't interrupt xneopsd functionality.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>