CA-413304: Restore VBD.unplug function to keep old functionality
This is a partial revert of 1a46f33be768. It retains the deactivate and
detach functions introduced but restores the original unplug function so
that the VBD_unplug atom is completely unchanged when
xenops_vbd_plug_unplug_legacy=true instead of running deactivate
followed by detach.

This will fix the S(Does_not_exist) Xenopsd errors we are seeing in
some VBD_unplug calls, until a fix for the split functions is found.

Signed-off-by: Steven Woods <steven.woods@cloud.com>