CA-413412: Fail to designate new master
The user attempted to designate a new master, but the operation failed.

The root cause is as follows:
After the new proposed master successfully sent the `commit_new_master`
API call to the old master, it attempted to send a `logout` request.
However, at this point, the old master was already rebooting its xapi
service, causing the `logout` to fail. As a result, the process of
designating the new master was marked as failed, and the status changed
to `broken`. In high-load environments, there can be a delay in sending
the logout request, increasing the likelihood that it is sent after the
old master has already started rebooting.

If `commit_new_master` has already been successful, the success of the
subsequent `logout` operation should not be considered critical.
Therefore, the solution is to ignore the result of the `logout` request
if `commit_new_master` was successful.

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>