CA-412420: Set vdi-type When Create snapshot on SMAPIv3 SR
During the SMAPIv3 migration prepare step, it will create the VDI on the destination based on the source.
The source retrieves vdi_info via copy_into_sr → find_vdi → SR.scan2 → sr_scan2_impl → COWVolume.ls,
and sr_ls extracts all vdi_info from the database.

However, when creating a snapshot with COWVolume.snapshot on V3, the vdi-type is not stored in the database,
which causes getting an empty vdi-type and the migration prepare step to fail and result the folllowing error:

`\'INTERNAL_ERROR\', \'Failure("Unknown tag/contents")\']]);S()]]]))]))]))])`

This change adds a step to set vdi-type from the original VDI to the snapshot VDI to prevent this issue.

Signed-off-by: Lunfan Zhang[Lunfan.Zhang] <Lunfan.Zhang@cloud.com>