CA-414418: Perf: save user validate result and apply to sessions
For all sessions created by external/AD users, session revalidate
will check whether the users are still acitve, and kick off the
session accordingly.

However, xapi check the user for every session. The problem here
is lots of session are created by only a few users. (for the case
of CVAD and ControlUP). This would cause lot of duplicated check
for the same user again and again, which is slow and waste lots
of resources.

To fix the issue, [(user_sid, check_result)] is defined for every
round of check. The check result is saved so later check for the
session with same user can just be applied.

Signed-off-by: Lin Liu <Lin.Liu01@cloud.com>