Update pool_metrics.user_agents
At the point of session login, check the user-agent in HTTP head
and record the user agents that are in a configured whitelist.
The user-agent format is <name>/<version> <comment>, parse the
name and version to update in pool_metrics.user_agents.
If the version is different with the stored user-agent in DB,
record the last-seen version.

Signed-off-by: Changlei Li <changlei.li@cloud.com>