CP-308859 Track user_agents
At the point of session login, check the user-agent in HTTP head
and track the user agents.
The user-agent format is <name>/<version> <comment>, parse the
name and version to record.
If the version is different from the stored user-agent, record
the last-seen version.
If the number of tracked user agents > max_num, remove the oldest
entry in the record.

Signed-off-by: Changlei Li <changlei.li@cloud.com>