CP-53858: Domain CPU ready RRD metric - runnable_any
Adding a new metric 'runnable_any' as % of time that at least one vCPU
of the domain is in the runnable state. It is the sum of the following 3
metrics:
- runstate_full_contention
- runstate_concurrency_hazard
- runstate_partial_contention

Naming it 'runnable_any' instead of 'runnable' is to resolve one problem
with rrd2csv: if we name it 'runnable', rrd2csv will select both
"runnable" and "runnable_vcpus" when the 'runnable' is used:

> rrd2csv AVERAGE:vm:<vm-uuid>:runnable
> timestamp, AVERAGE:vm:<vm-uuid>:runnable, AVERAGE:vm:<vm-uuid>:runnable_vcpus

This is because "runnable" is a prefix of "runnable_vcpus".

Naming it 'runnable_any', with rrd2csv:
* can select only 'runnable_any' if we use:
  rrd2csv AVERAGE:vm:<vm-uuid>:runnable_any
* can select only 'runnable_vcpus' if we use:
  rrd2csv AVERAGE:vm:<vm-uuid>:runnable_vcpus
* can select both 'runnable_any' and 'runnable_vcpus' if we use:
  rrd2csv AVERAGE:vm:<vm-uuid>:runnable

Naming it 'runnable_any' also makes it clearer as the metric is % of
time that at least one vCPU of the domain is in the runnable state.

Add max to "runnable_any" metric to follow the fix here:

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>
[Rebase with metric renaming and some fixes]
Signed-off-by: Gang Ji <gang.ji@cloud.com>