CA-412983: HA doesn't keep trying to start best-effort VM
The issue occurs in a scenario involving a HA-enabled pool. A VM with its
VM.ha_restart_priority set to best-effort is running on a host. The VM's disk
resides on the host's local storage. When the host goes down, the VM cannot be
restarted on other hosts due to the disk's local storage dependency. However,
after the host recovers and comes back online, the VM still does not
automatically start on the original host.

Expected behavior: The VM should automatically start on the original host once
it has recovered. Generally, this behavior should be applied to all non-agile
VMs.

Signed-off-by: Ming Lu <ming.lu@cloud.com>