XSI-1969 XSI-1973 use https for localhost migration
We observe an instability of localhost migration with NVidia SRIOV GPUs
since we starting to fall back to http (rather than https) for localhost
migration. Outside of development, localhost migration is used during
device model restart.

Use HTTPS for localhost migration. This seems to improve stability (but
is not the root cause).

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>