CA-414651 Fix device rename in PIF after RPU
After RPU from legacy network device name, the name in PIF
object should be updated. The new network device name is
in interface_tables.device_to_position_table. The position
of old device name in PIF object can be get by its associated
bridge and then correspond with the new name.
In this procedure, a exsiting bridge is used to get the position.
However, the bridge is used to update MTU. It is specially
handled when the device is under a bond. The bridge is actually
the parent bond bridge. This leads to no position is found for
old device name and then device name is not updated correctly.
Fix: use the associated bridge to find the position.

Signed-off-by: Changlei Li <changlei.li@cloud.com>