CP-309523: Make networkd_db utility return bridge MAC address
This is a simplified version of commit 846ce82ba3. The feature
introduced that commit is to remove the network device renaming
functionality. When the feature is merged, there will be 3 states of
releases:
1. without the feature.
2. with the feature but renaming is still working.
3. with the feature and renaming has been removed.

Originally, the upgrade path was intended to support transitions from
state 2 to state 3 only. However, it has become necessary to also
support upgrades directly from state 1 to state 3.

This commit is to enable the release in state 1 to upgrade to state 3.
The change is kept extremely small so it can be merged independently
without waiting for the full feature to be merged.

In details, during the upgrade, the host-installer can't know the
management interface as the eth<N> will not present in the environment
the host-installer running. So it needs the MAC address to find out the
management interface for setting up network.

Signed-off-by: Ming Lu <ming.lu@cloud.com>