CP-308800: Revert https_only sync logic
Currently, when xapi starts, it checks iptables port 80 and updates the
`https_only` based on the result. However, if xapi switches to using
firewalld in the future, the firewalld service configuration will not be
persisted. This means that after updating https_only, the firewalld
service configuration may revert to the default value after a host reboot.
This is not expected.

To resolve this issue, reverse the current behavior as: update the
firewalld service status based on the value of https_only during startup.

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>